requirements/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

# Database
psycopg2-binary==2.9.5

# Redis & Celery
redis==4.5.5
celery==5.3.1
django-celery-beat==2.8.1

# Channels (WebSockets)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# CORS
django-cors-headers==3.13.0

# Storage
django-storages==1.13.2
boto3==1.26.77

# Config & Env
python-decouple==3.7
python-dotenv==0.21.1

# Production server
gunicorn==20.1.0

# Error Tracking
sentry-sdk==1.29.2
